[untested] Integrate Maelle's visuals of dynamic footprints
Tests on iPhone will be run this afternoon by Maelle, I couldn't check how it works on the phone.
Adapted Youchun's code to not modify the data in PersonData but only at render time in the gameobject
Like that we can keep the original data to then display phones for example.