Colorable footprints. Fix Firebase bug. Clean files
Integrates Maelle's update using a 3D mesh for the footprints so that we can change the footprints colors on the fly.
Had to clean the code (again) to make it work without blinking

Fix a weird Firebase bug by doing the URL set just before using the database. If done too far apart or in a separate script, it might cause
		Firebase App initializing app __FIRAPP_DEFAULT (default 1).
to log twice, which makes Firebase unable to listen events (but it's still capable of sending data).

Cleaned some files related to old footprints methods

Detach Firebase listeners on Destroy